WASD for camera rotation around planet

+++ Assets/Scripts/Creature.cs
@@ -25,7 +25,7 @@ public abstract class Creature {
 	{
 		CurrentHealth = CurrentHealth - d;
         Debug.Log(CurrentHealth);
-        GameObject.GetComponentInChildren<Slider>().value = CurrentHealth / MaxHealth;
+        GameObject.GetComponentInChildren<Slider>().value = (float)CurrentHealth / (float)MaxHealth;
         if (CurrentHealth < 0) Die();
 	}
